Channel2Transform creates unique array names. Removed Cancels attribute from PowerPlateOp.EnterAverageConditional Sum_Expanded handles an empty array. Dirichlet.GetMean handles zero pseudo-counts. Moved TrueSkill tests into TrueSkillTests.
